refactor: re-usage internal functions
