+++ docs/plugins/hmr.md
@@ -114,7 +114,7 @@ const bootstrap = () => platformBrowserDynamic().bootstrapModule(AppModule);
 
 if (environment.hmr) {
   if (module[ 'hot' ]) {
-    hmrNgxsBootstrap(module, bootstrap);
+    hmrNgxsBootstrap(module, bootstrap).catch(err => console.log(err));
   } else {
     console.error('HMR is not enabled for webpack-dev-server!');
     console.log('Are you using the --hmr flag for ng serve?');
