+++ packages/hmr-plugin/src/internal/init.ts
@@ -1,0 +1,9 @@
+import { validateExistHmrStorage } from './common';
+
+/**
+ * Session storage: max size - 5 MB, in future need usage IndexDB (50MB)
+ * Session storage is used so that lazy modules can also be updated.
+ */
+export function hmrInit() {
+  validateExistHmrStorage();
+}
